Configures security to return only the entities the user has access to, based on their roles.
This is implemented on the service layer, though. The endpoint is not secured via Spring Security, which allows for an empty list if the user doesn't have access to any entity (more RESTful than a hard access denied).
This implementation also defines a specific role to POST to a resource. This could be combined with tenant-based access to allow a user to see its own resources without needing a role (maybe on a next commit).